requirements.txt
aiogram==2.23.1
fastapi==0.95.0
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
